+++ R/targets_misc.R
@@ -312,3 +312,110 @@ rename_samples <- function(phyloseq_component,
   return(new_pq_component)
 }
 ################################################################################
+
+
+################################################################################
+#' Match sample names from sam_data and fastq files
+#'
+#' @description
+#'
+#' <a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle">
+#' <img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>
+#'
+#' Useful for targets bioinformatic pipeline.
+#'
+#' @param path_sam_data (Required) Path to sample data file.
+#' @param sample_col_name (Required) The name of the column defining sample names in the sample data file.
+#' @param path_raw_seq (Required) Path to the folder containing fastq files
+#' @param pattern_remove_sam_data If not null, describe the pattern that will be deleted from
+#'   sam_data samples names.
+#' @param pattern_remove_fastq_files If not null, describe the pattern that will be deleted from
+#'   fastq files names.
+#' @param verbose (logical, default TRUE) If TRUE, print some additional messages.
+#' @param remove_undocumented_fastq_files (logical, default FALSE) If set to TRUE
+#'   fastq files not present in sam_data are removed from your folder. 
+#'   Keep a copy of those files somewhere before. 
+#' @param prefix Add a prefix to new samples names (ex. prefix = "samp") 
+#' @param ... Other parameters passed on to [utils::read.csv()] function.
+#' @return A list of two objects :
+#'   - $sam_names_matching is a tibble of corresponding samples names
+#'   - $sam_data is a sample data files including only matching sample names 
+#' @importFrom utils read.csv
+#' @export
+#' @author Adrien Taudière
+sam_data_matching_names <- function(
+  path_sam_data, 
+  sample_col_name, 
+  path_raw_seq, 
+  pattern_remove_sam_data = NULL,
+  pattern_remove_fastq_files = NULL,
+  verbose = TRUE,
+  remove_undocumented_fastq_files = FALSE,
+  prefix = NULL,
+  ...
+) {
+  sam_d <- read.csv(path_sam_data, ...)
+  names_sam_data <- sam_d[[sample_col_name]]
+  names_fastq_files_fullpath <- list.files(path_raw_seq)
+  names_fastq_files <- basename(names_fastq_files_fullpath)
+
+  if(!is.null(pattern_remove_sam_data)) {
+    names_sam_data_clean <- gsub(pattern_remove_sam_data, "", names_sam_data)
+    if (sum(duplicated(names_sam_data_clean)) > 0) {
+      stop(
+        "Their is duplicates in samples names from sam_data when removing pattern_remove_sam_data",
+        names_sam_data_clean[duplicated(names_sam_data_clean)]
+      )
+    }
+  }
+
+  if(!is.null(pattern_remove_fastq_files)) {
+    names_fastq_files_clean <- gsub(pattern_remove_fastq_files, "", names_fastq_files)
+  }
+
+  tib_sam_data <- tibble(
+    clean_sam = as.character(names_sam_data_clean), 
+    raw_sam = as.character(names_sam_data)
+  )
+
+  tib_fastq <- tibble(
+    clean_fastq = as.character(names_fastq_files_clean), 
+    raw_fastq = as.character(names_fastq_files),
+    raw_fastq_full_path = names_fastq_files_fullpath
+  )
+
+  tib_j <- full_join(tib_fastq, tib_sam_data, by = join_by(clean_fastq == clean_sam)) |>
+    rename(common_names = clean_fastq)
+
+  if(sum(is.na(tib_j$raw_fastq))>0){
+    message(sum(is.na(tib_j$raw_fastq)), " samples in sam_data files are not present in fastq_files")
+    if(verbose) {
+      warning(tib_j$raw_sam[is.na(tib_j$raw_fastq)], "not_matching_names_from_sam_data.txt")
+    }
+  }
+
+  if(sum(is.na(tib_j$raw_sam))>0){
+    message(sum(is.na(tib_j$raw_sam)), " samples in fastq files are not present in sam_data")
+    if(verbose) {
+      warning(tib_j$raw_fastq[is.na(tib_j$raw_sam)], "not_matching_names_from_fastq_files.txt")
+    }
+    if(remove_undocumented_fastq_files){
+      if(verbose) {
+        warning("Files ", tib_j$raw_fastq_full_path[is.na(tib_j$raw_sam)], "will be removed from folder", path_raw_seq)
+      }
+      unlink(tib_j$raw_fastq_full_path[is.na(tib_j$raw_sam)])
+    }
+  }
+  
+  if(!is.null(prefix)){
+    tib_j$common_names <- paste0(prefix, tib_j$common_names)
+  }
+
+  sam_d_new <- sam_d |>
+    dplyr::filter(.data[[sample_col_name]] %in% tib_j$raw_sam[!is.na(tib_j$raw_fastq)]) |>
+    dplyr::mutate("samples_names_common" = paste0(prefix, gsub(pattern_remove_sam_data, "", .data[[sample_col_name]]))) |>
+      relocate(samples_names_common)
+
+  return(list("sam_names_matching" = tib_j, "sam_data" = sam_d_new))
+}
+################################################################################